Remove incompatible `@types/d3` dependency
The `d3` dependency in the repo has a version specifier "^7.8.0", but
the latest published `@types/d3` package has version "7.4.3".

This prompts yarn to ask the user to select a `@types/d3` version
manually on every `yarn install`.

Since `d3` is only used in `ForceGraph.js` which is not a TS file,
the `@types/d3` dependency can be removed to resolve this issue.